# feat: Configurable EXPLAIN ANALYZE