--- a/statement_processing.go
+++ b/statement_processing.go
@@ -132,9 +132,6 @@ var (
 	explainColumnNames = []string{"ID", "Query_Execution_Plan <execution_method> (metadata, ...)"}
 	explainColumnAlign = []int{tablewriter.ALIGN_RIGHT, tablewriter.ALIGN_LEFT}
 
-	explainAnalyzeColumnNames = []string{"ID", "Query_Execution_Plan <execution_method> (metadata, ...)", "Rows_Returned", "Executions", "Total_Latency"}
-	explainAnalyzeColumnAlign = []int{tablewriter.ALIGN_RIGHT, tablewriter.ALIGN_LEFT, tablewriter.ALIGN_LEFT, tablewriter.ALIGN_LEFT, tablewriter.ALIGN_LEFT}
-
 	describeColumnNames = []string{"Column_Name", "Column_Type"}
 
 	// DDL needing special treatment