This PR implements configurable execution stats columns in `EXPLAIN ANALYZE`.

```
spanner> SET CLI_ANALYZE_COLUMNS='Rows:{{if ne .Rows.Total ""}}{{.Rows.Total}}{{end}},Scanned:{{.ScannedRows.Total}},Filtered:{{.FilteredRows.Total}}';

Empty set (0.00 sec)

spanner> EXPLAIN ANALYZE
         SELECT * FROM Singers
         JOIN Albums USING (SingerId)
         WHERE FirstName LIKE "M%c%";
+-----+--------------------------------------------------------------------------------+------+---------+----------+
| ID  | Query_Execution_Plan <execution_method> (metadata, ...)                        | Rows | Scanned | Filtered |
+-----+--------------------------------------------------------------------------------+------+---------+----------+
|   0 | Distributed Union on Singers <Row> (split_ranges_aligned: true)                |  864 |         |          |
|   1 | +- Local Distributed Union <Row>                                               |  864 |         |          |
|   2 |    +- Serialize Result <Row>                                                   |  864 |         |          |
|   3 |       +- Cross Apply <Row>                                                     |  864 |         |          |
|  *4 |          +- [Input] Filter Scan <Row> (seekable_key_size: 0)                   |      |         |          |
|   5 |          |  +- Table Scan on Singers <Row> (Full scan, scan_method: Automatic) |   27 |    1000 |      973 |
|  15 |          +- [Map] Local Distributed Union <Row>                                |  864 |         |          |
|  16 |             +- Filter Scan <Row> (seekable_key_size: 0)                        |      |         |          |
| *17 |                +- Table Scan on Albums <Row> (scan_method: Row)                |  864 |     864 |        0 |
+-----+--------------------------------------------------------------------------------+------+---------+----------+
Predicates(identified by ID):
  4: Residual Condition: ($FirstName LIKE 'M%c%')
 17: Seek Condition: ($SingerId_1 = $SingerId)
```

## Background

The `EXPLAIN ANALYZE` in spanner-cli was heavily influenced by the information displayed by the original Spanner plan visualizer.
https://github.com/cloudspannerecosystem/spanner-cli/issues/57

While the 3-column execution stats are sufficient for basic optimization, they only show a portion of the information that Spanner actually returns to the client. This new feature allows users to select the specific information they want to see, thereby reducing the need to use external tools like rendertree, spannerplanviz, or others for visualizing profiling data.

This feature is a simplified version of the template for [rendertree](https://github.com/apstndb/spannerplanviz/tree/main/cmd/rendertree)'s `--custom` flag.


## Breaking Changes

The default column name of `EXPLAIN ANALYZE` is changed to reduce line length.

### Before
```
+----+--------------------------------------------------------------------------+---------------+------------+---------------+
| ID | Query_Execution_Plan <execution_method> (metadata, ...)                  | Rows_Returned | Executions | Total_Latency |
+----+--------------------------------------------------------------------------+---------------+------------+---------------+
|  0 | Distributed Union on Singers <Row> (split_ranges_aligned: false)         | 5             | 1          | 1.08 msecs    |
|  1 | +- Local Distributed Union <Row>                                         | 5             | 1          | 1.05 msecs    |
|  2 |    +- Serialize Result <Row>                                             | 5             | 1          | 1.04 msecs    |
|  3 |       +- Table Scan on Singers <Row> (Full scan, scan_method: Automatic) | 5             | 1          | 1.03 msecs    |
+----+--------------------------------------------------------------------------+---------------+------------+---------------+
```
### After
```
+----+--------------------------------------------------------------------------+------+-------+---------------+
| ID | Query_Execution_Plan <execution_method> (metadata, ...)                  | Rows | Exec. | Total Latency |
+----+--------------------------------------------------------------------------+------+-------+---------------+
|  0 | Distributed Union on Singers <Row> (split_ranges_aligned: false)         |    5 |     1 |    0.66 msecs |
|  1 | +- Local Distributed Union <Row>                                         |    5 |     1 |    0.64 msecs |
|  2 |    +- Serialize Result <Row>                                             |    5 |     1 |    0.63 msecs |
|  3 |       +- Table Scan on Singers <Row> (Full scan, scan_method: Automatic) |    5 |     1 |    0.62 msecs |
+----+--------------------------------------------------------------------------+------+-------+---------------+
```